% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_accumulate_for_days.R
\name{accumulate_for_days}
\alias{accumulate_for_days}
\title{Delayed accumulation}
\usage{
accumulate_for_days(x, n)
}
\arguments{
\item{x}{vector to accumulate}

\item{n}{number of "times" each element count in the sum}
}
\value{
a vector of length length(x) + n -1
}
\description{
given a vector accumulate the sum supposing each elements count
for a given period of time.
}
\examples{

accumulate_for_days(c(1, 2, 3), 2) # c(1, 3, 5, 3)
accumulate_for_days(c(1, 2, 3), 3) # c(1, 3, 6, 5, 3)
accumulate_for_days(c(NA, 2, 3), 3) # c(0, 2, 5, 5, 3)
}
