% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19ita-package.R
\docType{package}
\name{covid19ita-package}
\alias{covid19ita}
\alias{covid19ita-package}
\title{covid19ita: Covid-19 - Italy}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Monitoring platform for Covid-19 infection
    diffusion in Italy.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UBESP-DCTV/covid19ita}
  \item Report bugs at \url{https://github.com/UBESP-DCTV/covid19ita/issues}
}

}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@unipd.it} (0000-0002-0520-7428)

Authors:
\itemize{
  \item Dario Gregori \email{dario.gregori@unipd.it} (0000-0001-7906-0580)
  \item Paola Berchialla \email{paola.berchialla@unito.it} (0000-0001-5835-5638)
  \item Danila Azzolina \email{danila.azzolina@uniupo.it} (0000-0002-8185-5742)
  \item Ilaria Prosepe \email{ilaria.prosepe@unipd.it}
  \item Giulia Lorenzoni \email{giulia.lorenzoni@unito.it} (0000-0003-1771-4686)
}

Other contributors:
\itemize{
  \item UBEP [copyright holder]
}

}
\keyword{internal}
