% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mort_data_comuni.R
\name{mort_data_comuni}
\alias{mort_data_comuni}
\title{Reschape comuni_settimana data tidyly}
\usage{
mort_data_comuni()
}
\value{
a tibble including regione, provincia,
  comune, settimana, classe di età, area (nord - or -
  centro, sud, isole), strata, year, n_death
}
\description{
Reschape comuni_settimana data tidyly
}
\examples{
mort_data_comuni()
}
