% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mort_data_veneto.R
\name{mort_data_veneto}
\alias{mort_data_veneto}
\title{Evaluate mortality variation}
\usage{
mort_data_veneto(type = c("sex", "age"))
}
\arguments{
\item{type}{(chr) "sex" or "age" stratification respectively}
}
\value{
a tibble including name, stratas, 2019 and 2020 overall death
  by strata, and the relative variation from 2019 to 2020
}
\description{
Evaluate mortality variation
}
\examples{
mort_data_veneto("sex")
mort_data_veneto("age")
}
