% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pull_region_w_pop.R
\name{pull_region_w_pop}
\alias{pull_region_w_pop}
\title{Extract region adding population}
\usage{
pull_region_w_pop(region)
}
\arguments{
\item{region}{(chr) name of the region to extract}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
From the official regional [dpc_covid19_ita_regioni] data, pull out
records from a sigle region, and add a column reporting that
region population
}
\examples{
covid19ita:::pull_region_w_pop("veneto")
}
