% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_raddopia.R
\name{raddopia}
\alias{raddopia}
\title{Raddopia un numero}
\usage{
raddopia(x)
}
\arguments{
\item{x}{(num) numero da raddoppiare}
}
\value{
Il numero raddopiato
}
\description{
Questas funzione è utile per raddopiare i numeri
}
\examples{
raddopia(2)
}
