% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(..., language = c("ita", "eng"))
}
\arguments{
\item{...}{A series of options to be used inside the app.}

\item{language}{(chr, default = "ita") the language of the application
i.e., "ita" for Italian or "eng" for English)}
}
\description{
Run the Shiny Application
}
