% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19ita-package.R
\docType{package}
\name{covid19ita-package}
\alias{covid19ita}
\alias{covid19ita-package}
\title{covid19ita: Covid-19 - Italy}
\description{
Monitoring platform for Covid-19 infection diffusion in Italy.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UBESP-DCTV/covid19ita}
  \item Report bugs at \url{https://github.com/UBESP-DCTV/covid19ita/issues}
}

}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@unipd.it}

Authors:
\itemize{
  \item Dario Gregori \email{dario.gregori@unipd.it}
}

Other contributors:
\itemize{
  \item UBEP [copyright holder]
}

}
\keyword{internal}
