% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_write_raw_rds.R
\name{write_raw_rds}
\alias{write_raw_rds}
\alias{read_data_raw}
\title{Write RDS for data-raw}
\usage{
write_raw_rds(data_name)

read_data_raw(data_name)
}
\arguments{
\item{data_name}{(chr) names of the pobject}
}
\value{
named logical vector (invisibly) with `TRUE`s for all the
  objects that are correctly saved, `FALSE`s for the other ones.
}
\description{
Given the names of some data, it save the corresponding objects in
the `here::here("data-raw/")` folder. Each object will be saved like
an independend `.rds` file with its own name as its filename.
}
\section{Functions}{
\itemize{
\item \code{read_data_raw}: read data-raw
}}

